Separating the server into files
*  Separate routes by group.
*  Running validation in a middleware.
